refactor: split Single Button Alert scene's protocols into separated folder and files
