+++ Pharus/Scenes/Alerts/Single Button Alert/Protocols/SingleButtonAlertFlow.swift
@@ -1,0 +1,12 @@
+//
+//  SingleButtonAlertFlow.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol SingleButtonAlertFlow {
+    func closeModal()
+}
