+++ Pharus/Scenes/Alerts/Single Button Alert/Protocols/SingleButtonAlertPresenterProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  SingleButtonAlertPresenterProtocol.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol SingleButtonAlertPresenterProtocol {
+    func closeModal()
+}
