+++ Pharus/Scenes/Alerts/Single Button Alert/Protocols/SingleButtonAlertViewDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  SingleButtonAlertViewDelegate.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol SingleButtonAlertViewDelegate: AnyObject {
+    func closeButtonTapped()
+}
